+++ riskgpt/workflows/__init__.py
@@ -3,9 +3,13 @@
 from .check_context_quality import check_context_quality
 from .external_context_enrichment import external_context_enrichment
 from .prepare_presentation_output import prepare_presentation_output
+from .risk_workflow import async_risk_workflow, fetch_relevant_documents, risk_workflow
 
 __all__ = [
     "prepare_presentation_output",
     "external_context_enrichment",
     "check_context_quality",
+    "risk_workflow",
+    "async_risk_workflow",
+    "fetch_relevant_documents",
 ]